% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute risk stratified analysis}
\usage{
execute(
  connectionDetails,
  analysisId,
  databaseSettings,
  maxDaysAtRisk,
  treatmentCohortId,
  comparatorCohortId,
  balanceThreads = 1,
  negativeControlThreads = 1,
  fitOutcomeModelsThreads = 1,
  createPsThreads = 1,
  saveDirectory = "results",
  removeSubjectsWithPriorOutcome = TRUE,
  customLimits = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{ConnectionDetails} created using
the \code{\link[DatabaseConnector]{createConnectionDetails}} function}

\item{analysisId}{Unique analysis identifier}

\item{databaseSettings}{Settings for locating the data in the database. Should
be created with the \code{\link[RiskStratifiedEstimation]{createDatabaseSettings}}
function}

\item{maxDaysAtRisk}{The maximum days at risk to be used for the analysis.
If patients have longer time at risk, they will be censored at the provided value}

\item{treatmentCohortId}{The ID of the treatment cohort in the table of the database
where exposures are stored}

\item{comparatorCohortId}{The ID of the comparator cohort in the table of the
database where exposures are stored}

\item{balanceThreads}{Number of parallel threads to be used for the evaluation
of covariate balance}

\item{negativeControlThreads}{Number of parallel threads to be used for the
negative control analyses}

\item{fitOutcomeModelsThreads}{Number of parallel threads to be used for
the estimation of the outcome models}

\item{createPsThreads}{Number of parallel threads to be used for fitting the
propensity score models}

\item{saveDirectory}{The directory where results will be stored}

\item{removeSubjectsWithPriorOutcome}{Should that have had the outcome prior
to the risk window start?}

\item{customLimits}{Custom limits used for risk stratification. If \code{NULL},
limits are c(0.01, 0.015)}
}
\description{
Executes a risk stratified analysis for the evaluation of
treatment effect heterogeneity of thiazides or thiazide-like diuretics compared
to ACE inhibitors.
}
