% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R
\name{runSensitivity}
\alias{runSensitivity}
\title{Run sensitivity analyses}
\usage{
runSensitivity(
  connectionDetails,
  databaseSettings,
  sensitivity,
  maxDaysAtRisk = 730,
  removeSubjectsWithPriorOutcome,
  useCustomLimits = TRUE,
  negativeControlThreads = 1,
  fitOutcomeModelsThreads = 1,
  createPsThreads = 1,
  balanceThreads = 1
)
}
\arguments{
\item{connectionDetails}{Database connection details. Should be created
with \code{\link[DatabaseConnector]{createConnectionDetails}}.}

\item{databaseSettings}{Information on the location of cohort tables. Should be
created with \code{\link[RiskStratifiedEstimation]{createDatabaseSettings}}}

\item{sensitivity}{A string describing the type of the sensitivity analysis.
Can one of "with_cvd" or "without_cvd".}

\item{maxDaysAtRisk}{Time point for prediction of absolute treatment effect.}

\item{removeSubjectsWithPriorOutcome}{Should prior outcomes be excluded?}

\item{useCustomLimits}{Should custom limits for risk stratification be used?}

\item{negativeControlThreads}{Nmber of parallel threads for the negative control analyses.}

\item{fitOutcomeModelsThreads}{Number of the parallel threads for fitting the outcome
models.}

\item{createPsThreads}{Number of threads for fitting the propensity score models.}

\item{balanceThreads}{Number of threads for evaluating covariate balance.}
}
\description{
Run sensitivity analyses for people with or without prior
             cardiovascular disease.
}
